Use the same link in the import and export modals
